Authorize occasions

+++ BengalApplication/app/controllers/occasions_controller.rb
@@ -1,16 +1,17 @@
 class OccasionsController < ApplicationController
+
   def index
-
-      @occasions = Occasion.all
-
+    @occasions = Occasion.all
   end
 
   def new
     @occasion = Occasion.new
+    authorize @occasion
   end
 
   def create
     occasion = Occasion.new(occasion_params)
+    authorize @occasion
     if occasion.save
       redirect_to occasions_path
     else
@@ -21,14 +22,17 @@ class OccasionsController < ApplicationController
 
   def show
     @occasion = Occasion.find(params[:id])
+    # authorize @occasion
   end
 
   def edit
     @occasion = Occasion.find(params[:id])
+    authorize @occasion
   end
 
   def update
     occasion = Occasion.find(params[:id])
+    authorize @occasion
     if occasion.update(occasion_params)
       redirect_to occasions_path
     else
@@ -39,6 +43,7 @@ class OccasionsController < ApplicationController
 
   def destroy
     occasion = Occasion.find(params[:id])
+    authorize occasion
     occasion.destroy
     redirect_to occasions_path
   end
